requirements.txt
streamlit>=1.35.0
streamlit-agraph>=0.0.45
neo4j>=5.10.0
openai>=1.0.0
python-dotenv>=1.0.0
duckduckgo-search>=5.0.0
watchdog>=4.0.0
